Expendable Items
Final Item Commit! CurrentMP stat was added to match CurrentHP since that was missing. Changes to classes to facilitate imports have been reverted with notes and comments. The only things not part of this mass update are:
1. Gold
2. Status Effects
3. Sound Effects (Listed as text strings, should be proper relationships)
4. Empowers (purposefully, as they are broken in vanilla)
5. Particle Effects (Listed as text strings, should be proper relationships)
Additionally, there's a potion spray checkbox, but we need to figure out the direction of data flow between the potion spray skill and this item, and figure out the best way to really connect them.